+++ musicdl/modules/sources/base.py
@@ -111,10 +111,47 @@ class BaseMusicClient():
         self.logger_handle.info(f'Finished searching music files using {self.source}. Search results have been saved to {work_dir}, valid items: {len(song_infos)}.', disable_print=self.disable_print)
         # return
         return song_infos
+    '''_youtubedownload'''
+    @usedownloadheaderscookies
+    def _youtubedownload(self, song_info: dict, request_overrides: dict = None, downloaded_song_infos: list = [], progress: Progress = None, 
+                         song_progress_id: int = 0, songs_progress_id: int = 0):
+        request_overrides = request_overrides or {}
+        try:
+            total_size, chunk_size, downloaded_size = int(song_info['download_url'].filesize), song_info.get('chunk_size', 1024 * 1024), 0
+            progress.update(song_progress_id, total=total_size)
+            save_path, same_name_file_idx = os.path.join(song_info['work_dir'], f"{song_info['song_name']}.{song_info['ext']}"), 1
+            while os.path.exists(save_path):
+                save_path = os.path.join(song_info['work_dir'], f"{song_info['song_name']}_{same_name_file_idx}.{song_info['ext']}")
+                same_name_file_idx += 1
+            with open(save_path, "wb") as fp:
+                for chunk in song_info['download_url'].iterchunks(chunk_size=chunk_size):
+                    if not chunk: continue
+                    fp.write(chunk)
+                    downloaded_size = downloaded_size + len(chunk)
+                    if total_size > 0:
+                        downloading_text = "%0.2fMB/%0.2fMB" % (downloaded_size / 1024 / 1024, total_size / 1024 / 1024)
+                    else:
+                        progress.update(song_progress_id, total=downloaded_size)
+                        downloading_text = "%0.2fMB/%0.2fMB" % (downloaded_size / 1024 / 1024, downloaded_size / 1024 / 1024)
+                    progress.advance(song_progress_id, len(chunk))
+                    progress.update(song_progress_id, description=f"{self.source}.download >>> {song_info['song_name']} (Downloading: {downloading_text})")
+            progress.advance(songs_progress_id, 1)
+            progress.update(song_progress_id, description=f"{self.source}.download >>> {song_info['song_name']} (Success)")
+            downloaded_song_info = copy.deepcopy(song_info)
+            downloaded_song_info['save_path'] = save_path
+            downloaded_song_infos.append(downloaded_song_info)
+        except Exception as err:
+            progress.update(song_progress_id, description=f"{self.source}.download >>> {song_info['song_name']} (Error: {err})")
+        return downloaded_song_infos
     '''_download'''
     @usedownloadheaderscookies
     def _download(self, song_info: dict, request_overrides: dict = None, downloaded_song_infos: list = [], progress: Progress = None, 
                   song_progress_id: int = 0, songs_progress_id: int = 0):
+        if not isinstance(song_info['download_url'], str) and song_info['source'] in ['YouTubeMusicClient']:
+            return self._youtubedownload(
+                song_info=song_info, request_overrides=request_overrides, downloaded_song_infos=downloaded_song_infos, progress=progress,
+                song_progress_id=song_progress_id, songs_progress_id=songs_progress_id,
+            )
         request_overrides = request_overrides or {}
         try:
             touchdir(song_info['work_dir'])